feat(mega-menu): dark mode "empty search" image
